+++ packages/amplify-graphql-model-transformer/src/__tests__/model-transformer.test.ts
@@ -1814,4 +1814,96 @@ describe('ModelTransformer:', () => {
       });
     });
   });
+
+  describe('remove null timestamp fields from input', () => {
+    it('updatedAt null', () => {
+      const validSchema = `
+        type UpdatedAtNull @model(timestamps: { updatedAt: null }) {
+            id: ID!
+            title: String!
+        }
+      `;
+
+      const out = testTransform({
+        schema: validSchema,
+        transformers: [new ModelTransformer()],
+      });
+      expect(out).toBeDefined();
+
+      validateModelSchema(parse(out.schema));
+      expect(out.schema).toMatchSnapshot();
+    });
+
+    it('createdAt null', () => {
+      const validSchema = `
+        type CreatedAtNull @model(timestamps: { createdAt: null }) {
+            id: ID!
+            title: String!
+        }
+      `;
+
+      const out = testTransform({
+        schema: validSchema,
+        transformers: [new ModelTransformer()],
+      });
+      expect(out).toBeDefined();
+
+      validateModelSchema(parse(out.schema));
+      expect(out.schema).toMatchSnapshot();
+    });
+
+    it('createdAt null and updatedAt null', () => {
+      const validSchema = `
+        type CreatedAtAndUpdatedAtNull @model(timestamps: { createdAt: null, updatedAt: null }) {
+            id: ID!
+            title: String!
+        }
+      `;
+
+      const out = testTransform({
+        schema: validSchema,
+        transformers: [new ModelTransformer()],
+      });
+      expect(out).toBeDefined();
+
+      validateModelSchema(parse(out.schema));
+      expect(out.schema).toMatchSnapshot();
+    });
+
+    it('timestamps null', () => {
+      const validSchema = `
+        type TimeStampsNull @model(timestamps: null) {
+            id: ID!
+            title: String!
+        }
+      `;
+
+      const out = testTransform({
+        schema: validSchema,
+        transformers: [new ModelTransformer()],
+      });
+      expect(out).toBeDefined();
+
+      validateModelSchema(parse(out.schema));
+      expect(out.schema).toMatchSnapshot();
+    });
+
+    it('custom createdAt and updatedAt null', () => {
+      const validSchema = `
+        type CustomCreatedAtAndUpdatedAtNull @model(timestamps: { createdAt: "createdOn", updatedAt: null }) {
+            id: ID!
+            title: String!
+        }
+      `;
+
+      const out = testTransform({
+        schema: validSchema,
+        transformers: [new ModelTransformer()],
+      });
+      expect(out).toBeDefined();
+
+      validateModelSchema(parse(out.schema));
+      expect(out.schema).toMatchSnapshot();
+    });
+  });
 });