+++ packages/amplify-data-construct/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.7.0...@aws-amplify/data-construct@1.7.1) (2024-04-11)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
 # [1.7.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.6.0...@aws-amplify/data-construct@1.7.0) (2024-03-28)
 
 ### Features