+++ packages/amplify-graphql-default-value-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-default-value-transformer@2.3.2...@aws-amplify/graphql-default-value-transformer@2.3.3) (2024-04-11)
+
+**Note:** Version bump only for package @aws-amplify/graphql-default-value-transformer
+
 ## [2.3.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-default-value-transformer@2.3.1...@aws-amplify/graphql-default-value-transformer@2.3.2) (2024-03-28)
 
 **Note:** Version bump only for package @aws-amplify/graphql-default-value-transformer